ref: better API abstraction
