+++ Core/Interface/APIs/TranslationAPI.py
@@ -1,0 +1,13 @@
+from abc import abstractmethod, ABC
+
+from Core.Interface.APIs.API import API
+
+
+class TranslationAPI(ABC, API):
+    @abstractmethod
+    def detect_lang(self):
+        pass
+
+    @abstractmethod
+    def translate(self):
+        pass
