+++ Infrastructure/Clients/Client.py
@@ -1,5 +1,4 @@
 from abc import ABC
-
 from requests import Session, Request
