+++ Infrastructure/Clients/GoogleClient.py
@@ -6,7 +6,9 @@ class GoogleClient(Client):
     def __init__(self, service: str):
         self.base_url = 'https://www.googleapis.com/auth/'
         self.service = service
-        super().__init__(Request())
+        super().__init__(
+            Request()
+        )
 
     @property
     def scopes(self) -> list[str]:
